+++ src/observable.ts
@@ -35,7 +35,7 @@ import { assertEventStream, assertFunction, assertNoArguments } from "./internal
 import skip from "./skip";
 import map from "./map";
 import flatMapConcat from "./flatmapconcat";
-import { sampledByE, sampledByP, sampleP } from "./sample";
+import { sampledBy, sampleP } from "./sample";
 import { filter } from "./filter";
 import { and, not, or } from "./boolean";
 import flatMapFirst from "./flatmapfirst";
@@ -72,7 +72,7 @@ import { diff, Differ } from "./diff";
 import { flatScan } from "./flatscan";
 import { holdWhen } from "./holdwhen";
 import { zip } from "./zip";
-import decode from "./decode";
+import decode, { DecodedValueOf } from "./decode";
 import { firstToPromise, toPromise } from "./topromise";
 import { more } from "./reply"
 import { withLatestFromE, withLatestFromP } from "./withlatestfrom";
@@ -108,6 +108,7 @@ export abstract class Observable<V> {
     this.desc = desc
     this.initialDesc = desc
   }
+
   /**
 Creates a Property that indicates whether
 `observable` is awaiting `otherObservable`, i.e. has produced a value after the latest
@@ -163,7 +164,7 @@ events from `other`. This means too that events from `other`,
 occurring before the end of this observable will not be included in the result
 stream/property.
    */
-  abstract concat(other: Observable<V>): Observable<V>
+  abstract concat<V2>(other: Observable<V2>): Observable<V | V2>
   /**
 Throttles stream/property by given amount
 of milliseconds, but so that event is only emitted after the given
@@ -218,8 +219,10 @@ The return value of [`decode`](#decode) is always a [`Property`](property.html).
 
    */
 
-  decode(cases: any): Property<any> {
-    return decode(this, cases)
+  //decode<T extends Record<any, any>>(src: Observable<keyof T>, cases: T): Property<DecodedValueOf<T>>
+
+  decode<T extends Record<any, any>>(cases: T): Property<DecodedValueOf<T>> {
+    return decode(<any>this, cases)
   }
   /**
 Delays the stream/property by given amount of milliseconds. Does not delay the initial value of a [`Property`](property.html).
@@ -461,7 +464,10 @@ Bacon.sequentially(2, events)
 ```
 
    */
+  abstract groupBy(keyF: Function1<V, string>, limitF?: GroupTransformer<V, V>): Observable<EventStream<V>>
+
   abstract groupBy<V2>(keyF: Function1<V, string>, limitF: GroupTransformer<V, V2>): Observable<EventStream<V2>>
+
   /**
 Pauses and buffers the event stream if last event in valve is truthy.
 All buffered events are released when valve becomes falsy.
@@ -604,7 +610,21 @@ Only applicable for observables with arrays as values.
 
    @param {Observable<V2>} sampler
    */
-  abstract sampledBy(sampler: Observable<any>): Observable<V>
+
+  /**
+   Creates an EventStream/Property by sampling this
+   stream/property value at each event from the `sampler` stream. The result
+   will contain the sampled value at each event in the source stream.
+
+   @param {Observable<V2>} sampler
+   */
+  sampledBy(sampler: EventStream<any>): EventStream<V>
+  sampledBy(sampler: Property<any>): Property<V>
+  sampledBy(sampler: Observable<any>): Observable<V>
+  sampledBy(sampler: Observable<any>): Observable<V> {
+    return sampledBy(this, sampler, arguments[1]) // TODO: combinator
+  }
+
   /**
 Scans stream/property with given seed value and
 accumulator function, resulting to a Property. For example, you might
@@ -969,8 +989,8 @@ export class Property<V> extends Observable<V> {
    occurring before the end of this property will not be included in the result
    stream/property.
    */
-  concat(other: Observable<V>): Property<V> {
-    return addPropertyInitValueToStream<V>(this, this.changes().concat(other))
+  concat<V2>(other: Observable<V2>): Property<V | V2> {
+    return addPropertyInitValueToStream<V | V2>(this as Property<V | V2>, this.changes().concat(other))
   }
 
   /** @hidden */
@@ -1087,8 +1107,12 @@ export class Property<V> extends Observable<V> {
    ```
 
    */
+  groupBy(keyF: Function1<V, string>, limitF?: GroupTransformer<V, V>): Property<EventStream<V>>
+
+  groupBy<V2>(keyF: Function1<V, string>, limitF: GroupTransformer<V, V2>): Property<EventStream<V2>>
+
   groupBy<V2>(keyF: Function1<V, string>, limitF: GroupTransformer<V, V2>): Property<EventStream<V2>> {
-    return <any>groupBy(this, keyF, limitF)
+    return <any>groupBy<V, V2>(this, keyF, limitF)
   }
 
   /**
@@ -1120,18 +1144,6 @@ export class Property<V> extends Observable<V> {
    */
   sample(interval: number): EventStream<V> {
     return sampleP(this, interval)
-  }
-
-  /**
-   Creates an EventStream by sampling this
-   stream/property value at each event from the `sampler` stream. The result
-   `EventStream` will contain the sampled value at each event in the source
-   stream.
-
-   @param {Observable<V2>} sampler
-   */
-  sampledBy(sampler: Observable<any>): Observable<V> {
-    return sampledByP(this, sampler, arguments[1])
   }
 
   /**
@@ -1333,7 +1345,7 @@ export class EventStream<V> extends Observable<V> {
    occurring before the end of this observable will not be included in the result
    stream/property.
    */
-  concat(other: Observable<V>, options?: EventStreamOptions): EventStream<V> {
+  concat<V2>(other: Observable<V2>, options?: EventStreamOptions): EventStream<V | V2> {
     return concatE(this, other, options)
   }
   /** @hidden */
@@ -1446,6 +1458,10 @@ export class EventStream<V> extends Observable<V> {
    ```
 
    */
+  groupBy(keyF: Function1<V, string>, limitF?: GroupTransformer<V, V>): EventStream<EventStream<V>>
+
+  groupBy<V2>(keyF: Function1<V, string>, limitF: GroupTransformer<V, V2>): EventStream<EventStream<V2>>
+
   groupBy<V2>(keyF: Function1<V, string>, limitF: GroupTransformer<V, V2>): EventStream<EventStream<V2>> {
     return <any>groupBy(this, keyF, limitF)
   }
@@ -1461,27 +1477,15 @@ export class EventStream<V> extends Observable<V> {
   /**
    Merges two streams into one stream that delivers events from both
    */
-  merge(other: EventStream<V>): EventStream<V> {
+  merge<V2>(other: EventStream<V2>): EventStream<V  | V2> {
     assertEventStream(other)
-    return mergeAll<V>(this, other).withDesc(new Desc(this, "merge", [other]));
+    return mergeAll<V | V2>(this as EventStream<V | V2>, other as EventStream<V | V2>).withDesc(new Desc(this, "merge", [other]));
   }
 
   /**
    Returns a stream/property that inverts boolean values (using `!`)
    */
   not(): EventStream<boolean> {return <any>not(this) }
-
-  /**
-   Creates an EventStream by sampling this
-   stream/property value at each event from the `sampler` stream. The result
-   `EventStream` will contain the sampled value at each event in the source
-   stream.
-
-   @param {Observable<V2>} sampler
-   */
-  sampledBy(sampler: Observable<any>): Observable<V> {
-    return sampledByE(this, sampler, arguments[1])
-  }
 
   /**
    Adds a starting value to the stream/property, i.e. concats a