+++ test/groupby.ts
@@ -2,12 +2,12 @@ import * as Bacon from "..";
 
 import { expectStreamEvents, series, unstable, semiunstable } from "./util/SpecHelper";
 
-function flattenAndConcat<V>(obs: Bacon.EventStream<Bacon.Observable<V>>) {
+function flattenAndConcat<V>(obs: Bacon.EventStream<Bacon.EventStream<V>>) {
   return obs.flatMap((obs: Bacon.Observable<V>) =>
     obs.fold([], (xs: any, x: any) => xs.concat(x)))
 }
 
-function flattenAndMerge<V>(obs: Bacon.EventStream<Bacon.Observable<V>>) {
+function flattenAndMerge<V>(obs: Bacon.EventStream<Bacon.EventStream<V>>) {
   return obs.flatMap(Bacon._.id);
 }
 
@@ -23,17 +23,17 @@ function takeWhileInclusive<V>(obs: Bacon.EventStream<V>, f: (v: V) => boolean) 
   })
 }
 
-function toString<A>(x: number) {
+function toString(x: number) {
   return x.toString()
 }
 
 describe("EventStream.groupBy", function() {
   describe("without limiting function (legacy support)", function() {
     expectStreamEvents(
-      () => flattenAndConcat(series(2, [1,2,2,3,1,2,2,3]).groupBy(toString, <any>undefined)),
+      () => flattenAndConcat(series(2, [1,2,2,3,1,2,2,3]).groupBy(toString)),
       [[1,1],[2,2,2,2],[3,3]], unstable);
     return expectStreamEvents(
-      () => flattenAndMerge(series(2, [1,2,2,3,1,2,2,3]).groupBy(toString, <any>undefined)),
+      () => flattenAndMerge(series(2, [1,2,2,3,1,2,2,3]).groupBy(toString)),
       [1,2,2,3,1,2,2,3], semiunstable);
   });
   describe("with limiting function", function() {