+++ types/argumentstoobservables.d.ts
@@ -1,3 +1,0 @@
-import Observable from "./observable";
-export declare function argumentsToObservables<T>(args: (Observable<T> | Observable<T>[] | T)[]): Observable<T>[];
-export declare function argumentsToObservablesAndFunction(args: any): [Observable<any>[], Function];