+++ types/asyncwrapsubscribe.d.ts
@@ -1,2 +1,0 @@
-import { Subscribe } from "./types";
-export default function asyncWrapSubscribe<V>(obs: any, subscribe: Subscribe<V>): Subscribe<V>;